+++ CHANGELOG.md
@@ -34,6 +34,7 @@ Types of changes: `Added`, `Changed`, `Deprecate`, `Removed`, `Fixed`, `Secruity
 - Added command to create a new model
 - Added web cli
 - Added all CRUD actions for user with UserController
+- Added get() and set() function to base model
 
 ### Changed
 - Improved the autoloader to work for the new folder structure of version 2